Fix FRB codegen for RGBA dimensions
flutter_rust_bridge 1.80.1 does not support primitive lists containing usize. Return RGBA frame dimensions as Vec<i32> instead of Vec<usize> so bridge generation can parse the new API.

+++ src/flutter_ffi.rs
@@ -2253,7 +2253,7 @@ pub fn session_get_rgba_size(session_id: SessionID, display: usize) -> SyncRetur
 pub fn session_get_rgba_dimension(
     session_id: SessionID,
     display: usize,
-) -> SyncReturn<Vec<usize>> {
+) -> SyncReturn<Vec<i32>> {
     SyncReturn(super::flutter::session_get_rgba_dimension(
         session_id, display,
     ))
